Add calls to g++ to salsa queries
